Extend contract-package pattern to all content verticals
Ship @voyantjs/accommodations-contracts, @voyantjs/products-contracts,
@voyantjs/extras-contracts, and @voyantjs/charters-contracts as zod-only
packages that own each vertical's <vertical>/v1 content schema, version
constant, types, and validator, so external consumers (Voyant Connect,
adapter authors, the Admin API SDK) can validate payloads without the
framework runtime. The runtime packages re-export from them so existing
@voyantjs/<vertical>/content-shape import paths are unchanged; the
mergeOverlaysInto<Vertical>Content overlay composition stays in the runtime.

Add ADR-0002 codifying the *-contracts boundary — what is pure versus
runtime, the runtime -> contracts dependency direction, and the
re-export-stub recipe — so future verticals and the Admin API SDK follow one
documented pattern.

Relates to #1400, #1411.
