+++ packages/extras-contracts/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./content-shape.js"
